Pass mapDispatch test in Landing page, move files to respective directories
