+++ src/Components/App/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Header from '../../Containers/Header/Header';
-import LandingPage from '../LandingPage/LandingPage';
+import LandingPage from '../../Containers/LandingPage/LandingPage';
 import Login from '../../Containers/Login/Login';
 import ResearchContainer from '../ResearchContainer/ResearchContainer';
 import './App.scss';
